Parametrize PoW to make tests fatser
