+++ chain/consensus/tspow/mine.go
@@ -2,19 +2,14 @@ package tspow
 
 import (
 	"context"
-	"crypto/rand"
+	"fmt"
 	"time"
 
-	"golang.org/x/xerrors"
-
 	"github.com/filecoin-project/go-address"
-	"github.com/filecoin-project/go-state-types/big"
-	"github.com/filecoin-project/lotus/chain/consensus/platform/logging"
-
 	lapi "github.com/filecoin-project/lotus/api"
 	"github.com/filecoin-project/lotus/api/v1api"
-	"github.com/filecoin-project/lotus/chain/consensus/common"
 	param "github.com/filecoin-project/lotus/chain/consensus/common/params"
+	"github.com/filecoin-project/lotus/chain/consensus/platform/logging"
 	"github.com/filecoin-project/lotus/chain/types"
 )
 
@@ -23,7 +18,7 @@ func Mine(ctx context.Context, miner address.Address, api v1api.FullNode) error 
 
 	head, err := api.ChainHead(ctx)
 	if err != nil {
-		return xerrors.Errorf("getting head: %w", err)
+		return fmt.Errorf("getting head: %w", err)
 	}
 
 	log.Info("starting PoW mining on @", head.Height())
@@ -47,7 +42,7 @@ func Mine(ctx context.Context, miner address.Address, api v1api.FullNode) error 
 			lbr := base.Height() + 1 - DiffLookback(base.Height())
 			lbts, err := api.ChainGetTipSetByHeight(ctx, lbr, base.Key())
 			if err != nil {
-				return xerrors.Errorf("failed to get lookback tipset+1: %w", err)
+				return fmt.Errorf("failed to get lookback tipset+1: %w", err)
 			}
 
 			expDiff = Difficulty(base, lbts)
@@ -114,41 +109,4 @@ func Mine(ctx context.Context, miner address.Address, api v1api.FullNode) error 
 
 		log.Info("PoW mined a block! ", bh.Cid())
 	}
-}
-
-func (tsp *TSPoW) CreateBlock(ctx context.Context, w lapi.Wallet, bt *lapi.BlockTemplate) (*types.FullBlock, error) {
-	b, err := common.PrepareBlockForSignature(ctx, tsp.sm, bt)
-	if err != nil {
-		return nil, err
-	}
-	next := b.Header
-
-	tgt := big.Zero()
-	tgt.SetBytes(next.Ticket.VRFProof)
-
-	bestH := *next
-	for i := 0; i < 10000; i++ {
-		next.ElectionProof = &types.ElectionProof{
-			VRFProof: []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
-		}
-		rand.Read(next.ElectionProof.VRFProof) //nolint:errcheck
-		if work(&bestH).LessThan(work(next)) {
-			bestH = *next
-			if work(next).GreaterThanEqual(tgt) {
-				break
-			}
-		}
-	}
-	next = &bestH
-
-	if work(next).LessThan(tgt) {
-		return nil, nil
-	}
-
-	err = common.SignBlock(ctx, w, b)
-	if err != nil {
-		return nil, err
-	}
-
-	return b, nil
 }
