+++ chain/consensus/tspow/pow_test.go
@@ -5,9 +5,9 @@ import (
 	"math/rand"
 	"testing"
 
-	"github.com/filecoin-project/go-state-types/big"
 	"github.com/ipfs/go-cid"
 
+	"github.com/filecoin-project/go-state-types/big"
 	"github.com/filecoin-project/lotus/chain/actors/builtin"
 	"github.com/filecoin-project/lotus/chain/types"
 )
